moved hnse serilizer functions to a separate header added hnsw ctor for the serializer(required also new ctor for VecSimIndexAbstract) and factory functions to call this ctor, added VecSimAlgo_INVALID, removed friend class SerializerTest_HNSWSerialzer_Test, added serilized index v1 to unit data and the v1 only loads it
