+++ src/VecSim/vec_sim.cpp
@@ -125,8 +125,9 @@ extern "C" size_t VecSimIndex_EstimateInitialSize(const VecSimParams *params) {
         return HNSWFactory::EstimateInitialSize(&params->hnswParams);
     case VecSimAlgo_BF:
         return BruteForceFactory::EstimateInitialSize(&params->bfParams);
+    default:
+        return -1;
     }
-    return -1;
 }
 
 extern "C" int VecSimIndex_AddVector(VecSimIndex *index, const void *blob, size_t id) {
@@ -153,8 +154,9 @@ extern "C" size_t VecSimIndex_EstimateElementSize(const VecSimParams *params) {
         return HNSWFactory::EstimateElementSize(&params->hnswParams);
     case VecSimAlgo_BF:
         return BruteForceFactory::EstimateElementSize(&params->bfParams);
+    default:
+        return -1;
     }
-    return -1;
 }
 
 extern "C" void VecSim_Normalize(void *blob, size_t dim, VecSimType type) {
